Moves function `setColumnDefsAndConstraints` to `CreateStatement.cpp`.
