+++ src/sql/statements.cpp
@@ -138,19 +138,6 @@ CreateStatement::~CreateStatement() {
   }
 }
 
-void CreateStatement::setColumnDefsAndConstraints(std::vector<TableElement*>* tableElements) {
-  columns = new std::vector<ColumnDefinition*>();
-  tableConstraints = new std::vector<TableConstraint*>();
-
-  for (auto tableElem : *tableElements) {
-    if (auto* colDef = dynamic_cast<ColumnDefinition*>(tableElem)) {
-      columns->emplace_back(colDef);
-    } else if (auto* tableConstraint = dynamic_cast<TableConstraint*>(tableElem)) {
-      tableConstraints->emplace_back(tableConstraint);
-    }
-  }
-}
-
 // DeleteStatement
 DeleteStatement::DeleteStatement() : SQLStatement(kStmtDelete), schema(nullptr), tableName(nullptr), expr(nullptr){};
